+++ os/axvisor/src/main.rs
@@ -31,6 +31,7 @@ use ax_std as _;
 
 mod banner;
 mod config;
+mod host_devices;
 mod manager;
 #[cfg(target_arch = "riscv64")]
 mod platform_irq;